# backend/internal/restapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI Code Review API
  version: 1.0.0
  description: |
    Read API for the dashboard, plus two admin actions. Authenticate with
    `Authorization: Bearer <token>`; tokens and roles come from `API_TOKENS`.
    This identifies the calling *server*, not an end user.
    Every operation carries `x-required-role` (none | viewer | admin); a test
    keeps it equal to what the server enforces.
servers:
  - url: /
security:
  - bearer: []
paths:
...
  /api/v1/me:
    get:
      summary: Role of the calling token and the user it belongs to
      description: |
        `user` is null for a token that is not linked to a person (the dashboard's
        shared read token, or an admin token without a user id).
      x-required-role: viewer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [role, user]
                properties:
                  role: { type: string, enum: [viewer, author, senior, lead, admin] }
                  user:
                    nullable: true
                    type: object
                    required: [id, name]
                    properties: { id: { type: integer, format: int64 }, name: { type: string } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/overview:
    get:
      summary: Totals and a daily series for the last N UTC days
      x-required-role: viewer
      parameters:
        - { $ref: "#/components/parameters/days" }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/Overview" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/repositories:
    get:
      summary: List repositories
      x-required-role: viewer
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 200 }, description: Case-insensitive match on workspace/slug }
        - { name: review_enabled, in: query, schema: { type: boolean } }
        - { $ref: "#/components/parameters/limit" }
        - { $ref: "#/components/parameters/offset" }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/RepositoryPage" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/repositories/{id}:
    parameters:
      - { $ref: "#/components/parameters/id" }
    get:
      summary: One repository
      x-required-role: viewer
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/Repository" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: Turn automatic review of a repository on or off
      description: Off by default, because enabling it sends code to an LLM.
      x-required-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [review_enabled], additionalProperties: false, properties: { review_enabled: { type: boolean } } }
      responses:
        "200":
          description: Updated repository
          content: { application/json: { schema: { $ref: "#/components/schemas/Repository" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/commits:
    get:
      summary: List commits, newest first
      description: Keyset pagination; pass `next_cursor` back as `cursor`. Each commit shows the score of its latest review.
      x-required-role: viewer
      parameters:
        - { name: repo_id, in: query, schema: { type: integer, format: int64, minimum: 1 } }
        - { name: author_id, in: query, schema: { type: integer, format: int64, minimum: 1 } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/ReviewStatus" } }
        - { name: branch, in: query, schema: { type: string, maxLength: 200 }, description: Exact branch name }
        - { name: q, in: query, schema: { type: string, maxLength: 200 }, description: Case-insensitive match on message or hash }
        - { name: since, in: query, schema: { type: string, format: date-time }, description: committed_at >= since }
        - { name: until, in: query, schema: { type: string, format: date-time }, description: committed_at < until }
        - { name: cursor, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/limit" }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/CommitPage" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/commits/{id}:
    parameters:
      - { $ref: "#/components/parameters/id" }
    get:
      summary: One commit with its latest review, findings and suggestions
      x-required-role: viewer
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/CommitDetail" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/commits/{id}/rereview:
    parameters:
      - { $ref: "#/components/parameters/id" }
    post:
      summary: Queue the commit for another review
      description: |
        Sets the commit back to `pending` and enqueues a job. 409 when the
        review is running, the repository has review disabled, or it is a
        merge commit. The previous review stays visible until a new one is saved.
      x-required-role: admin
      responses:
        "202":
          description: Queued
          content: { application/json: { schema: { type: object, required: [status], properties: { status: { type: string, enum: [queued] } } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/findings/{id}/fixed:
    parameters:
      - { $ref: "#/components/parameters/id" }
    post:
      summary: The author says the finding is fixed
      description: |
        Only the author of the commit or pull request (the user the token belongs
        to) may do this, and only while the finding is `open` and the review is
        not closed. A senior, lead or admin may stand in when the change has no
        linked author, and an admin may do it for anyone (an admin token needs no
        user; what it does is recorded without a name). The optional note says
        what was done.
      x-required-role: author
      requestBody:
        required: false
        content: { application/json: { schema: { type: object, properties: { note: { type: string, maxLength: 2000 } } } } }
      responses:
        "200":
          description: Done
          content: { application/json: { schema: { type: object, required: [status], properties: { status: { type: string, enum: [ok] } } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/findings/{id}/reopen:
    parameters:
      - { $ref: "#/components/parameters/id" }
    post:
      summary: Send a fixed or dismissed finding back to the author
      description: |
        A senior, lead or admin who is not the author (an admin token without a
        user is allowed too). `note` is required (what is still wrong). On a closed review this reopens the review too.
      x-required-role: senior
      requestBody:
        required: false
        content: { application/json: { schema: { type: object, properties: { note: { type: string, maxLength: 2000 } } } } }
      responses:
        "200":
          description: Done
          content: { application/json: { schema: { type: object, required: [status], properties: { status: { type: string, enum: [ok] } } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/findings/{id}/dismiss:
    parameters:
      - { $ref: "#/components/parameters/id" }
    post:
      summary: Say the finding is not a real problem
      description: |
        A senior, lead or admin who is not the author. `note` is required (why).
        409 on a closed review.
      x-required-role: senior
      requestBody:
        required: false
        content: { application/json: { schema: { type: object, properties: { note: { type: string, maxLength: 2000 } } } } }
      responses:
        "200":
          description: Done
          content: { application/json: { schema: { type: object, required: [status], properties: { status: { type: string, enum: [ok] } } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/reviews/{id}/close:
    parameters:
      - { $ref: "#/components/parameters/id" }
    post:
      summary: Close the review
      description: |
        A senior, lead or admin who is not the author. 409 while any finding is
        still `open`, or when the review is already closed.
      x-required-role: senior
      requestBody:
        required: false
        content: { application/json: { schema: { type: object, properties: { note: { type: string, maxLength: 2000 } } } } }
      responses:
        "200":
          description: Done
          content: { application/json: { schema: { type: object, required: [status], properties: { status: { type: string, enum: [ok] } } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/pull-requests:
    get:
      summary: List pull requests, most recently updated first
      description: |
        Keyset pagination; pass `next_cursor` back as `cursor`. A pull request is
        reviewed as one diff (all its commits together), separately from its
        commits' own reviews; `score` and `findings` are from its latest review.
        Only pull requests of repositories the poller reads are listed.
      x-required-role: viewer
      parameters:
        - { name: repo_id, in: query, schema: { type: integer, format: int64, minimum: 1 } }
        - { name: author_id, in: query, schema: { type: integer, format: int64, minimum: 1 } }
        - { name: state, in: query, schema: { $ref: "#/components/schemas/PullRequestState" } }
        - { name: review_status, in: query, schema: { $ref: "#/components/schemas/ReviewStatus" } }
        - { name: q, in: query, schema: { type: string, maxLength: 200 }, description: Case-insensitive match on title, source branch or number }
        - { name: cursor, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/limit" }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/PullRequestPage" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/pull-requests/{id}:
    parameters:
      - { $ref: "#/components/parameters/id" }
    get:
      summary: One pull request with its latest review, findings and suggestions
      x-required-role: viewer
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/PullRequestDetail" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/pull-requests/{id}/rereview:
    parameters:
      - { $ref: "#/components/parameters/id" }
    post:
      summary: Queue the pull request for another review
      description: |
        Sets it back to `pending` and enqueues a job. 409 when a review is
        running, the repository has review disabled, or the pull request is no
        longer open. The previous review stays visible until a new one is saved.
      x-required-role: admin
      responses:
        "202":
          description: Queued
          content: { application/json: { schema: { type: object, required: [status], properties: { status: { type: string, enum: [queued] } } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/users:
    get:
      summary: List people with review statistics for the last N days
      description: "`email` is only returned to admin tokens."
      x-required-role: viewer
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 200 }, description: Case-insensitive match on name or nickname }
        - { name: sort, in: query, schema: { type: string, enum: [name, commits, score], default: name } }
        - { $ref: "#/components/parameters/days" }
        - { $ref: "#/components/parameters/limit" }
        - { $ref: "#/components/parameters/offset" }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/UserPage" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/users/{id}:
    parameters:
      - { $ref: "#/components/parameters/id" }
    get:
      summary: One person with a weekly trend
      description: "`email` is only returned to admin tokens. Weeks without commits are omitted from `trend`."
      x-required-role: viewer
      parameters:
        - { $ref: "#/components/parameters/days" }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/UserDetail" } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
  parameters:
    id: { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
    limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    offset: { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
    days: { name: days, in: query, schema: { type: integer, minimum: 1, maximum: 365, default: 30 }, description: Window in UTC days, today included }
  responses:
    BadRequest: { description: Invalid parameter, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Unauthorized: { description: Missing or invalid token, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Forbidden: { description: Token role is not allowed, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    NotFound: { description: Not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    Conflict: { description: State does not allow it, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties: { code: { type: string }, message: { type: string } }
    ReviewStatus: { type: string, enum: [pending, running, done, skipped, failed] }
...
    Overview:
      type: object
      properties:
        days: { type: integer }
        since: { type: string, format: date-time }
        commits: { type: integer }
        commits_by_status: { type: object, additionalProperties: { type: integer } }
        reviewed: { type: integer, description: Commits that have a review }
        avg_score: { type: number, nullable: true }
        usage:
          type: object
          description: |
            Review runs created in the window, re-reviews included (cost is per run).
            `measured_runs` counts runs that reported usage; mock runs report none and add nothing to the sums.
            Tokens are everything sent to / returned by the model (cached input included);
            cost is the provider-reported figure, in USD.
          properties:
            runs: { type: integer }
            measured_runs: { type: integer }
            tokens_in: { type: integer, format: int64 }
            tokens_out: { type: integer, format: int64 }
            cost_usd: { type: number }
            avg_cost_usd: { type: number, nullable: true, description: Mean over measured runs }
        findings_by_severity: { type: object, additionalProperties: { type: integer }, description: Findings of each commit's latest review }
        queue_waiting: { type: integer }
        queue_running: { type: integer }
        active_repositories: { type: integer }
        active_authors: { type: integer }
        enabled_repositories: { type: integer }
        total_repositories: { type: integer }
        fix:
          type: object
          description: Fix workflow of the latest reviews of the commits in the window
          properties:
            open_findings: { type: integer, description: Waiting for the author }
            fixed_findings: { type: integer, description: Fixed by the author, review not closed yet }
            reviews_ready_to_close: { type: integer, description: Every finding fixed or dismissed, waiting for a senior, lead or admin }
            reviews_closed: { type: integer }
        series:
          type: array
          items:
            type: object
            properties:
              day: { type: string, format: date }
              commits: { type: integer }
              reviewed: { type: integer }
              avg_score: { type: number, nullable: true }
              cost_usd: { type: number, description: Cost of review runs created that day }
    Repository:
      type: object
      properties:
        id: { type: integer, format: int64 }
        workspace: { type: string }
        project_key: { type: string }
        project_name: { type: string }
        slug: { type: string }
        name: { type: string }
        full_name: { type: string }
        main_language: { type: string, nullable: true }
        default_branch: { type: string, nullable: true }
        review_enabled: { type: boolean }
        commits: { type: integer }
        reviewed: { type: integer }
        avg_score: { type: number, nullable: true }
        last_commit_at: { type: string, format: date-time, nullable: true }
    RepositoryPage:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Repository" } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    CommitSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        hash: { type: string }
        subject: { type: string, description: First line of the message }
        repository:
          type: object
          properties: { id: { type: integer, format: int64 }, full_name: { type: string } }
        author:
          type: object
          description: "`id` is null for authors known only from the git author string."
          properties:
            id: { type: integer, format: int64, nullable: true }
            name: { type: string }
            avatar_url: { type: string, nullable: true }
        branch: { type: string, nullable: true }
        committed_at: { type: string, format: date-time }
        review_status: { $ref: "#/components/schemas/ReviewStatus" }
        review_skip_reason: { type: string, nullable: true }
        files_changed: { type: integer, nullable: true }
        additions: { type: integer, nullable: true }
        deletions: { type: integer, nullable: true }
        is_merge: { type: boolean }
        score: { type: integer, nullable: true, minimum: 0, maximum: 100 }
        findings: { type: integer }
        reviewed_at: { type: string, format: date-time, nullable: true }
        open_findings: { type: integer, description: Findings of the latest review that are still open (the author has to fix them) }
        review_closed: { type: boolean, description: A senior, lead or admin closed the latest review }
    PullRequestState:
      type: string
      enum: [OPEN, MERGED, DECLINED, SUPERSEDED, DELETED]
      description: Bitbucket's state; `DELETED` means Bitbucket no longer knows the pull request.
    PullRequestSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        number: { type: integer, description: The pull request id in Bitbucket }
        title: { type: string }
        repository:
          type: object
          properties: { id: { type: integer, format: int64 }, full_name: { type: string } }
        author:
          type: object
          properties:
            id: { type: integer, format: int64, nullable: true }
            name: { type: string }
            avatar_url: { type: string, nullable: true }
        source_branch: { type: string, nullable: true }
        destination_branch: { type: string, nullable: true }
        state: { $ref: "#/components/schemas/PullRequestState" }
        review_status: { $ref: "#/components/schemas/ReviewStatus" }
        review_skip_reason: { type: string, nullable: true }
        files_changed: { type: integer, nullable: true }
        additions: { type: integer, nullable: true }
        deletions: { type: integer, nullable: true }
        score: { type: integer, nullable: true, minimum: 0, maximum: 100 }
        findings: { type: integer }
        reviewed_at: { type: string, format: date-time, nullable: true }
        open_findings: { type: integer, description: Findings of the latest review that are still open (the author has to fix them) }
        review_closed: { type: boolean, description: A senior, lead or admin closed the latest review }
        review_outdated: { type: boolean, description: The branch has new commits since the latest review }
        updated_at: { type: string, format: date-time }
    PullRequestPage:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/PullRequestSummary" } }
        next_cursor: { type: string, nullable: true }
    PullRequestDetail:
      allOf:
        - { $ref: "#/components/schemas/PullRequestSummary" }
        - type: object
          properties:
            description: { type: string }
            reviews_count: { type: integer }
            review: { allOf: [{ $ref: "#/components/schemas/Review" }], nullable: true }
    CommitPage:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/CommitSummary" } }
        next_cursor: { type: string, nullable: true }
...
    UserRef:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
    Review:
      type: object
      properties:
        id: { type: integer, format: int64 }
        model: { type: string, description: "`mock` marks results that did not come from a real model" }
        prompt_version: { type: string }
        score: { type: integer, nullable: true }
        summary: { type: string }
        duration_ms: { type: integer, nullable: true }
        tokens_in: { type: integer, nullable: true, description: Null when the reviewer reported no usage }
        tokens_out: { type: integer, nullable: true }
        cost_usd: { type: number, nullable: true }
        created_at: { type: string, format: date-time }
        closed:
          nullable: true
          description: Set once a senior, lead or admin closed the review
          type: object
          properties:
            at: { type: string, format: date-time }
            by: { allOf: [{ $ref: "#/components/schemas/UserRef" }], nullable: true }
            note: { type: string, nullable: true }
        findings: { type: array, items: { $ref: "#/components/schemas/Finding" } }
    CommitDetail:
      allOf:
        - { $ref: "#/components/schemas/CommitSummary" }
        - type: object
          properties:
            message: { type: string }
            reviews_count: { type: integer }
            review: { allOf: [{ $ref: "#/components/schemas/Review" }], nullable: true }
...
    UserPage:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/User" } }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    UserDetail:
      allOf:
        - { $ref: "#/components/schemas/User" }
        - type: object
          properties:
            days: { type: integer }
            trend:
              type: array
              items:
                type: object
                properties:
                  week: { type: string, format: date, description: Monday of the ISO week (UTC) }
                  commits: { type: integer }
                  reviewed: { type: integer }
                  avg_score: { type: number, nullable: true }